fix: major issue at decorators
Previously, the implementation was limiting the usage of each decorator
by one, because it was assuming that a single instance of a decorator
would be equal to a single usage of it in a function. However that
assumption is *WRONG* because a decorator gets instantiated once and
then applied multiple times.

This fixes it by moving the decorator metadata to the function object
instead of keeping it in the single tenant instance of the decorator
object.